Moved filters into views_helper.py, deleted now obsolete filters.py, moved (wrongly placed by us) functions from run_v2.py to run.py
