+++ protzilla/run.py
@@ -4,9 +4,14 @@ import logging
 import threading
 import traceback
 
+import os
+import shutil
+import datetime
+
 import protzilla.constants.paths as paths
 from protzilla.steps import Messages, Output, Plots, Step
 from protzilla.utilities import format_trace
+from protzilla.disk_operator import DiskOperator, YamlOperator
 
 
 def get_available_run_names() -> list[str]:
@@ -19,6 +24,64 @@ def get_available_run_names() -> list[str]:
         if not directory.name.startswith(".")
     ]
 
+def get_available_runinfo() -> tuple[list[dict[str, str | list[str]]], list[dict[str, str | list[str]]], set[str]]:
+    if not paths.RUNS_PATH.exists():
+        return []
+    runs = []
+    runs_favourited = []
+    all_tags = set()
+    for directory in paths.RUNS_PATH.iterdir():
+        if directory.name.startswith("."):
+            continue
+        name = directory.name
+        creation_time = directory.stat().st_ctime
+        modification_time = directory.stat().st_mtime
+
+        disk_operator = DiskOperator(name, "dummy_workflow_name")
+        directory_path = os.path.join(paths.RUNS_PATH, name)
+        run_yaml_path = os.path.join(directory_path, "run.yaml")
+        step_manager = disk_operator.read_run(run_yaml_path)
+        steps = step_manager.all_steps
+        step_names = []
+        for step in steps:
+            step_names.append(step.display_name)
+
+        tags = []
+        metadata_yaml_path = os.path.join(directory_path, "metadata.yaml")
+        if os.path.isfile(metadata_yaml_path):
+            yaml_operator = YamlOperator()
+            metadata = yaml_operator.read(metadata_yaml_path)
+            tags = metadata.get("tags")
+        
+        for tag  in tags:
+            all_tags.add(tag)
+
+        run = { 
+            "run_name": name,
+            "creation_date": datetime.datetime.fromtimestamp(creation_time).strftime("%d %m %Y"), #TODO: reutrn the pure datetime, convert in html)
+            "modification_date": datetime.datetime.fromtimestamp(modification_time).strftime("%d %m %Y"),
+            "memory_mode": step_manager.df_mode,
+            "run_steps" : step_names,
+            "favourite_status" : step_manager.favourite,
+            "run_tags": tags
+            }
+        
+        if step_manager.favourite: 
+            runs_favourited.append(run)
+        else:
+            runs.append(run)
+
+    for run in runs + runs_favourited:
+        possible_tags = list(all_tags - run["run_tags"])
+        run["addable_tags"] = possible_tags
+
+    return (runs, runs_favourited, all_tags)
+
+def delete_run_folder(run_name) -> None:
+    path = os.path.join(paths.RUNS_PATH, run_name)
+
+    if os.path.isdir(path):
+        shutil.rmtree(path)
 
 class Run:
     class ErrorHandlingContextManager:
